# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  psychologist: 'psychologist',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MoodEntry {
  id: string;
  userId: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  moodScore: number;
  emotions: string[];
  note?: string | null;
  createdAt: string;
}

export interface MoodEntryOrNull {
  entry: MoodEntry | null;
}

export interface MoodEntriesResponse {
  entries: MoodEntry[];
  total: number;
}

export interface CreateMoodEntryRequest {
  /**
     * @minimum 1
     * @maximum 10
     */
  moodScore: number;
  emotions: string[];
  note?: string | null;
}

export interface AffirmationResponse {
  text: string;
  date: string;
}

export type GetMoodEntriesParams = {
limit?: number;
offset?: number;
};