# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HOLA! Life Buddy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: mood
    description: Mood tracking
  - name: user
    description: User profile

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/users/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/mood:
    get:
      operationId: getMoodEntries
      tags: [mood]
      summary: Get mood entries for current user
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntriesResponse"
    post:
      operationId: createMoodEntry
      tags: [mood]
      summary: Log a mood entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoodEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"

  /v1/mood/today:
    get:
      operationId: getTodayMood
      tags: [mood]
      summary: Get today's mood entry for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntryOrNull"

  /v1/affirmation:
    get:
      operationId: getDailyAffirmation
      tags: [user]
      summary: Get today's daily affirmation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffirmationResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, psychologist, admin]
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    MoodEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        moodScore:
          type: integer
          minimum: 1
          maximum: 10
        emotions:
          type: array
          items:
            type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - moodScore
        - emotions
        - createdAt

    MoodEntryOrNull:
      type: object
      properties:
        entry:
          allOf:
            - $ref: "#/components/schemas/MoodEntry"
          nullable: true
      required:
        - entry

    MoodEntriesResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/MoodEntry"
        total:
          type: integer
      required:
        - entries
        - total

    CreateMoodEntryRequest:
      type: object
      properties:
        moodScore:
          type: integer
          minimum: 1
          maximum: 10
        emotions:
          type: array
          items:
            type: string
        note:
          type: string
          nullable: true
      required:
        - moodScore
        - emotions

    AffirmationResponse:
      type: object
      properties:
        text:
          type: string
        date:
          type: string
      required:
        - text
        - date