# lib/api-zod/src/generated/types/createMoodEntryRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMoodEntryRequest {
  /**
     * @minimum 1
     * @maximum 10
     */
  moodScore: number;
  emotions: string[];
  note?: string | null;
}