# lib/api-zod/src/generated/types/moodEntriesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */
import type { MoodEntry } from './moodEntry';

export interface MoodEntriesResponse {
  entries: MoodEntry[];
  total: number;
}