# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLA! Life Buddy API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}